custom-user-none/requirements.txt
asgiref==3.5.2
boto3==1.24.69
botocore==1.27.69
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-environ==0.9.0
django-resized==1.0.2
django-ses==3.1.2
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tzdata==2022.2
urllib3==1.26.12
